Update 20. Valid Parentheses.java
I not only made a change to make the code run in O(n) but also I used stack and kept the code clean so it is easy to read and simple to maintain. Please feel free to ask if you have any issue. 
Thanks for trusting and assign me to this issue, I am looking forward to collaborate with you more in future.